# docs/api/v1/swagger.yaml
definitions:
  entity.Task:
    properties:
      activeAt:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  v1.requestTask:
    properties:
      activeAt:
        type: string
      title:
        type: string
    required:
    - activeAt
    - title
    type: object
  v1.resp:
    properties:
      id:
        type: string
    type: object
info:
  contact: {}
paths:
  /tasks:
    get:
      description: Get a list of tasks based on the provided status
      parameters:
      - description: Status of the tasks (active, done)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Task'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: List tasks
    post:
      consumes:
      - application/json
      description: Create a new task with the provided title and activeAt date
      parameters:
      - description: Task details
        in: body
        name: requestTask
        required: true
        schema:
          $ref: '#/definitions/v1.requestTask'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.resp'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create task
  /tasks/{id}:
    delete:
      description: Delete an existing task based on its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete task
    put:
      consumes:
      - application/json
      description: Update the details of an existing task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task details
        in: body
        name: requestTask
        required: true
        schema:
          $ref: '#/definitions/v1.requestTask'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update task
  /tasks/{id}/done:
    put:
      description: Mark an existing task as done based on its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Mark task as done
swagger: "2.0"